Allow integer slider to support 64 bit ints (#10972)
* allow integer slider to support 64 bit ints

* fix failing tests, deserialize xml based integer slider nodes as int64 sliders

* update usage of BuildInt

* restore IntegerSlider for 32 bit type to prevent API break

* attempt to deserialize legacy integer slider node as 64 bit int slider

* code fix

* retain behaviour of legacy int 32 bit sliders

* address review comments

* update tests

* review comments, update test

* change output type to long

* fix output port type for 64 bit int slider

* fix regressions after updating output tooltip of JSON based integer slider nodes

* hide old 32 bit int slider class, update icon resources

* fix pending failing tests

* review comments

* cleanup commented code

+++ src/Libraries/CoreNodeModels/Input/BaseTypes.cs
@@ -737,7 +737,7 @@ namespace CoreNodeModels.Input
             public AssociativeNode GetAstNode(Dictionary<string, AssociativeNode> idLookup)
             {
                 if (Math.Floor(_d) == _d)
-                    return AstFactory.BuildIntNode((int)_d);
+                    return AstFactory.BuildIntNode((long)_d);
                 return AstFactory.BuildDoubleNode(_d);
             }
         }
